Only faff with Sublime Text on Darwin.

+++ .zshrc
@@ -25,8 +25,10 @@ setopt autolist
 
 if which rbenv > /dev/null; then eval "$(rbenv init -)"; fi
 
-if [ ! -L ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User ]; then
-	echo "Installing Sublime Text config."
-	rm -rf ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User
-	ln -sf ~/.config/Sublime\ Text ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User
+if [ $OS = Darwin ]; then
+	if [ ! -L ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User ]; then
+		echo "Installing Sublime Text config."
+		rm -rf ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User
+		ln -sf ~/.config/Sublime\ Text ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User
+	fi
 fi
